microservice/requirements.txt
pylint-django
black
django==4.1.5
djangorestframework==3.14.0
django-concurrency==2.4
drf_yasg==1.21.4
psycopg2-binary